more class moving to more well-defined package names
