+++ plan-app/src/main/java/sparkengine/plan/app/runner/PlanRunner.java
@@ -17,9 +17,9 @@ import sparkengine.plan.model.mapper.sql.SqlPlanMapper;
 import sparkengine.plan.model.plan.mapper.PlanMapper;
 import sparkengine.plan.model.plan.mapper.PlanMapperException;
 import sparkengine.plan.model.plan.mapper.PlanMappers;
-import sparkengine.plan.runtime.PipelineRunnersFactory;
-import sparkengine.plan.runtime.PipelineRunnersFactoryException;
-import sparkengine.plan.runtime.builder.ModelPipelineRunnersFactory;
+import sparkengine.plan.runtime.runner.PipelineRunnersFactory;
+import sparkengine.plan.runtime.runner.PipelineRunnersFactoryException;
+import sparkengine.plan.runtime.builder.runner.ModelPipelineRunnersFactory;
 import sparkengine.plan.runtime.datasetconsumer.DatasetConsumerException;
 import sparkengine.spark.sql.logicalplan.PlanExplorerException;
 import sparkengine.spark.sql.logicalplan.tablelist.TableListExplorer;
