+++ plan-model-builder/src/test/resources/referenceMapperTest/absoluteComponentReference.yaml
