+++ plan-model-builder/src/test/resources/referenceMapperTest/absoluteComponentReference_component2.yaml
