+++ plan-model-builder/src/test/resources/referenceMapperTest/nestedPlan_providedRef_component1.yaml
