+++ plan-model-builder/src/test/resources/referenceMapperTest/nestedPlan_wrapperComponentName_wrapper_wrapper_wrapper_wrapper_wrapper.yaml
