+++ plan-runtime/src/main/java/sparkengine/plan/runtime/runner/PipelineRunnersFactoryException.java
@@ -1,4 +1,4 @@
-package sparkengine.plan.runtime;
+package sparkengine.plan.runtime.runner;
 
 public class PipelineRunnersFactoryException extends Exception {
